Update engine version to 0.1.4
 * Updated engine version to 0.1.4
 * Fixed CMakeLists.txt so that the engine compiles with warnings enabled
 * Fixed game and engine version CMake defines
 * Fixed several naming issues related to shadow warnings
 * Adjusted OpenGL initialization output
 * Updated SDL2 initialization to use SDL_INIT_EVERYTHING
 * Temporarily fixed messenger handling (again)
        ! Somehow the MessageContents shared_ptrs are becoming empty, I may
          change them to unique pointers if possible

+++ bee/ui/objects/optionbox.hpp
@@ -25,12 +25,12 @@ class ObjUIOptionBox : public ObjUIElement {
 		void push_option(bee::Instance*, const std::string&, bool, std::function<void (bee::Instance*, bool)>);
 		void pop_option(bee::Instance*);
 
-		std::string get_option(bee::Instance*, int);
-		bool get_option_state(bee::Instance*, int);
+		std::string get_option(bee::Instance*, size_t);
+		bool get_option_state(bee::Instance*, size_t);
 		int get_option_at(bee::Instance*, int, int);
 		std::vector<int> get_selected_options(bee::Instance*);
 
-		void set_option_state(bee::Instance*, int, bool);
+		void set_option_state(bee::Instance*, size_t, bool);
 };
 
 #endif // BEE_UI_OBJ_OPTIONBOX_H
